refactor: create utils common module
