+++ src/controllers/badges/linkAccounts.ts
@@ -6,7 +6,7 @@ import { getSession } from "next-auth/client"
 import { checkUserSignature, linkAccounts } from "src/core/badges"
 import { dbConnect } from "src/utils/backend/database"
 import logger from "src/utils/backend/logger"
-import capitalize from "src/utils/common/capitalize"
+import { capitalize } from "src/utils/common"
 
 export default async function linkAccountsController(req: NextApiRequest, res: NextApiResponse) {
     if (req.method !== "PUT") {
