+++ src/core/contracts/ReputationBadge/exists.ts
@@ -1,7 +1,7 @@
 import { OAuthProvider } from "@interrep/reputation"
 import { ContractName } from "src/config"
 import getBackendContractInstance from "src/utils/backend/getBackendContractInstance"
-import getContractAddress from "src/utils/common/getContractAddress"
+import { getContractAddress } from "src/utils/common"
 
 export default async function exists(tokenId: string, provider: OAuthProvider): Promise<boolean> {
     const contractAddress = getContractAddress(ContractName.REPUTATION_BADGE, provider)
