+++ src/core/contracts/ReputationBadge/safeMint.ts
@@ -2,7 +2,7 @@ import { OAuthProvider } from "@interrep/reputation"
 import { ContractReceipt } from "ethers"
 import { ContractName } from "src/config"
 import getBackendContractInstance from "src/utils/backend/getBackendContractInstance"
-import getContractAddress from "src/utils/common/getContractAddress"
+import { getContractAddress } from "src/utils/common"
 
 export default async function safeMint(to: string, tokenId: string, provider: OAuthProvider): Promise<ContractReceipt> {
     const contractAddress = getContractAddress(ContractName.REPUTATION_BADGE, provider)
