+++ src/pages/badges.tsx
@@ -12,10 +12,7 @@ import { ContractName, currentNetwork } from "src/config"
 import EthereumWalletContext, { EthereumWalletContextType } from "src/context/EthereumWalletContext"
 import createUserMessage from "src/core/badges/createUserMessage"
 import useInterRepAPI from "src/hooks/useInterRepAPI"
-import capitalize from "src/utils/common/capitalize"
-import delay from "src/utils/common/delay"
-import getContractAddress from "src/utils/common/getContractAddress"
-import getContractInstance from "src/utils/common/getContractInstance"
+import { capitalize, delay, getContractAddress, getContractInstance } from "src/utils/common"
 import { ExplorerDataType, getExplorerLink } from "src/utils/frontend/getExplorerLink"
 
 export default function Badges(): JSX.Element {
