+++ src/utils/backend/getBackendContractInstance.ts
@@ -1,7 +1,7 @@
 import { Contract } from "ethers"
 import { ContractName } from "src/config"
-import getContractInstance from "src/utils/common/getContractInstance"
 import getSigner from "src/utils/backend/getSigner"
+import { getContractInstance } from "src/utils/common"
 
 export default async function getBackendContractInstance(
     contractName: ContractName,
