+++ src/utils/common/capitalize.ts
@@ -1,3 +1,8 @@
-export default function capitalize(s: string) {
-    return s.replace(/\w\S*/g, (w) => w.replace(/^\w/, (c) => c.toUpperCase()))
+/**
+ * Capitalizes a string.
+ * @param string The string to capitalize.
+ * @returns The capitalized string.
+ */
+export default function capitalize(string: string) {
+    return string.replace(/\w\S*/g, (w) => w.replace(/^\w/, (c) => c.toUpperCase()))
 }
