+++ src/utils/common/getContractEvents.ts
@@ -1,5 +1,12 @@
 import { Contract, Event } from "ethers"
 
+/**
+ * Return the events of a contract.
+ * @param contractInstance The contract instance.
+ * @param eventName The name of the event.
+ * @param eventParameters A list of possible parameters to be used.
+ * @returns A list of events.
+ */
 export default function getContractEvents(
     contractInstance: Contract,
     eventName: string,
