+++ src/utils/common/index.ts
@@ -1,0 +1,8 @@
+import delay from "./delay"
+import capitalize from "./capitalize"
+import getContractAddress from "./getContractAddress"
+import getContractEvents from "./getContractEvents"
+import getContractInstance from "./getContractInstance"
+import getNetworkFullName from "./getNetworkFullName"
+
+export { delay, capitalize, getContractAddress, getContractEvents, getContractInstance, getNetworkFullName }
