--- src/utils/common/stringToBigNumber.ts
@@ -1,5 +1,0 @@
-import { BigNumber } from "ethers"
-
-export default function stringToBigNumber(s: string): BigNumber {
-    return BigNumber.from(s)
-}
